# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/password/reset/request": {
      "post": {
        "description": "Sends a letter with an OTP code to reset the password",
        "operationId": "PasswordManagementController_resetPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Reset password mail has been successfully sent",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponseApiModel"
                    },
                    {
                      "properties": {
                        "status": {
                          "type": "string",
                          "default": "success"
                        },
                        "data": {
                          "$ref": "#/components/schemas/OtpTicketApiModel"
                        }
                      },
                      "required": [
                        "status",
                        "data"
                      ]
                    }
                  ]
                }
              }
            }
          },
          "403": {
            "description": "User status policy failed",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponseApiModel"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "status": {
                          "type": "string",
                          "enum": [
                            "error"
                          ],
                          "default": "error"
                        },
                        "error": {
                          "type": "object",
                          "properties": {
                            "type": {
                              "type": "string",
                              "enum": [
                                "access-control.authorization.blocked",
                                "access-control.authorization.forbidden"
                              ]
                            },
                            "message": {
                              "type": "string",
                              "default": "Forbidden."
                            },
                            "statusCode": {
                              "type": "number",
                              "enum": [
                                403
                              ]
                            },
                            "details": {
                              "oneOf": [
                                {
                                  "type": "object",
                                  "properties": {
                                    "reason": {
                                      "type": "string"
                                    }
                                  },
                                  "required": [
                                    "reason"
                                  ]
                                },
                                {
                                  "type": "object",
                                  "properties": {
                                    "reason": {
                                      "type": "string"
                                    }
                                  },
                                  "required": [
                                    "reason"
                                  ]
                                }
                              ]
                            }
                          },
                          "required": [
                            "type",
                            "statusCode"
                          ]
                        }
                      },
                      "required": [
                        "status",
                        "error"
                      ]
                    }
                  ]
                },
                "examples": {
                  "access-control.authorization.blocked_1": {
                    "summary": "User account is blocked",
                    "value": {
                      "status": "error",
                      "error": {
                        "type": "access-control.authorization.blocked",
                        "message": "Authorization failed.",
                        "statusCode": 403,
                        "details": {
                          "reason": "Your account has been blocked. Please contact support."
                        }
                      }
                    }
                  },
                  "access-control.authorization.forbidden_2": {
                    "summary": "Forbidden resource (fallback)",
                    "value": {
                      "status": "error",
                      "error": {
                        "type": "access-control.authorization.forbidden",
                        "message": "Authorization failed.",
                        "statusCode": 403,
                        "details": {
                          "reason": "Access denied. Please contact support."
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponseApiModel"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "status": {
                          "type": "string",
                          "enum": [
                            "error"
                          ],
                          "default": "error"
                        },
                        "error": {
                          "type": "object",
                          "properties": {
                            "type": {
                              "type": "string",
                              "default": "app.entity-not-found"
                            },
                            "message": {
                              "type": "string",
                              "default": "Entity not found."
                            },
                            "statusCode": {
                              "type": "number",
                              "default": 404
                            },
                            "details": {
                              "type": "object",
                              "properties": {
                                "entityType": {
                                  "type": "string"
                                },
                                "identity": {
                                  "type": "object",
                                  "additionalProperties": {
                                    "nullable": true,
                                    "oneOf": [
                                      {
                                        "type": "object",
                                        "title": "Object",
                                        "description": "Any JSON",
                                        "additionalProperties": true
                                      },
                                      {
                                        "type": "array",
                                        "items": {},
                                        "title": "Array",
                                        "description": "Any array"
                                      },
                                      {
                                        "type": "string",
                                        "title": "String",
                                        "description": "Any string"
                                      },
                                      {
                                        "type": "number",
                                        "title": "Number",
                                        "description": "Any number"
                                      },
                                      {
                                        "type": "boolean",
                                        "title": "String",
                                        "description": "True or false"
                                      }
                                    ]
                                  }
                                }
                              },
                              "required": [
                                "entityType"
                              ]
                            }
                          },
                          "required": [
                            "type",
                            "statusCode"
                          ]
                        }
                      },
                      "required": [
                        "status",
                        "error"
                      ]
                    }
                  ]
                },
                "example": {
                  "status": "error",
                  "error": {
                    "type": "app.entity-not-found",
                    "message": "Entity not found.",
                    "statusCode": 404,
                    "details": {
                      "entityType": "User",
                      "identity": {
                        "id": "6wvHiPEGR5X3wTPtTkjEhS"
                      }
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponseApiModel"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "status": {
                          "type": "string",
                          "enum": [
                            "error"
                          ],
                          "default": "error"
                        },
                        "error": {
                          "type": "object",
                          "properties": {
                            "type": {
                              "type": "string",
                              "default": "validation"
                            },
                            "message": {
                              "type": "string",
                              "default": "Validation failed."
                            },
                            "statusCode": {
                              "type": "number",
                              "default": 422
                            },
                            "details": {
                              "nullable": true,
                              "oneOf": [
                                {
                                  "type": "object",
                                  "title": "Object",
                                  "description": "Any JSON",
                                  "additionalProperties": true
                                },
                                {
                                  "type": "array",
                                  "items": {},
                                  "title": "Array",
                                  "description": "Any array"
                                },
                                {
                                  "type": "string",
                                  "title": "String",
                                  "description": "Any string"
                                },
                                {
                                  "type": "number",
                                  "title": "Number",
                                  "description": "Any number"
                                },
                                {
                                  "type": "boolean",
                                  "title": "String",
                                  "description": "True or false"
                                }
                              ]
                            }
                          },
                          "required": [
                            "type",
                            "statusCode"
                          ]
                        }
                      },
                      "required": [
                        "status",
                        "error"
                      ]
                    }
                  ]
                },
                "example": {
                  "status": "error",
                  "error": {
                    "type": "validation",
                    "message": "Validation failed.",
                    "statusCode": 422,
                    "details": {
                      "entries": [
                        {
                          "property": "email",
                          "value": "invalid-mail@mailcom",
                          "constraints": {
                            "isEmail": "Value must be an email."
                          }
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "429": {
            "description": "Rate limit",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponseApiModel"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "status": {
                          "type": "string",
                          "enum": [
                            "error"
                          ],
                          "default": "error"
                        },
                        "error": {
                          "type": "object",
                          "properties": {
                            "type": {
                              "type": "string",
                              "default": "rate-limit.mail"
                            },
                            "message": {
                              "type": "string",
                              "default": "Rate limit."
                            },
                            "statusCode": {
                              "type": "number",
                              "default": 429
                            },
                            "details": {
                              "type": "object",
                              "properties": {
                                "reason": {
                                  "type": "string"
                                },
                                "secondsLeft": {
                                  "type": "number"
                                }
                              },
                              "required": [
                                "reason",
                                "secondsLeft"
                              ]
                            }
                          },
                          "required": [
                            "type",
                            "statusCode"
                          ]
                        }
                      },
                      "required": [
                        "status",
                        "error"
                      ]
                    }
                  ]
                },
                "example": {
                  "status": "error",
                  "error": {
                    "type": "rate-limit.mail",
                    "message": "Rate limit.",
                    "statusCode": 429,
                    "details": {
                      "reason": "You have exceeded the rate limit. Please try again in 10 seconds.",
                      "secondsLeft": 10
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "The reset password mail could not be sent",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponseApiModel"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "status": {
                          "type": "string",
                          "enum": [
                            "error"
                          ],
                          "default": "error"
                        },
                        "error": {
                          "type": "object",
                          "properties": {
                            "type": {
                              "type": "string",
                              "default": "internal"
                            },
                            "message": {
                              "type": "string",
                              "default": "Internal server error."
                            },
                            "statusCode": {
                              "type": "number",
                              "default": 500
                            },
                            "details": {
                              "type": "object",
                              "properties": {
                                "reason": {
                                  "type": "string"
                                }
                              },
                              "required": [
                                "reason"
                              ]
                            }
                          },
                          "required": [
                            "type",
                            "statusCode"
                          ]
                        }
                      },
                      "required": [
                        "status",
                        "error"
                      ]
                    }
                  ]
                },
                "example": {
                  "status": "error",
                  "error": {
                    "type": "internal",
                    "message": "Internal server error.",
                    "statusCode": 500,
                    "details": {
                      "reason": "The reset password mail could not be sent."
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Reset password request",
        "tags": [
          "Password management"
        ]
      }
    },
...
  },
  "info": {
    "title": "Food Friend API",
    "description": "API documentation for the Food Friend app",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "Access token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "description": "Enter JWT token"
      }
    },
    "schemas": {
...
      "OtpTicketApiModel": {
        "type": "object",
        "properties": {
          "ticket": {
            "type": "string",
            "description": "OTP Ticket",
            "example": "LnT8BAUhhoJ2Y6MuB9AAZp"
          }
        },
        "required": [
          "ticket"
        ]
      },
...
      "HttpExceptionApiModel": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "description": "Exception type"
          },
          "statusCode": {
            "type": "number",
            "description": "HTTP status code"
          },
          "message": {
            "type": "string",
            "description": "Exception message"
          },
          "details": {
            "description": "Exception details",
            "oneOf": [
              {
                "type": "object",
                "title": "Object",
                "description": "Any JSON",
                "additionalProperties": true
              },
              {
                "type": "array",
                "items": {},
                "title": "Array",
                "description": "Any array"
              },
              {
                "type": "string",
                "title": "String",
                "description": "Any string"
              },
              {
                "type": "number",
                "title": "Number",
                "description": "Any number"
              },
              {
                "type": "boolean",
                "title": "String",
                "description": "True or false"
              }
            ],
            "nullable": true
          }
        },
        "required": [
          "type",
          "statusCode",
          "message",
          "details"
        ]
      },
      "ErrorResponseApiModel": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "description": "Response status",
            "enum": [
              "error"
            ]
          },
          "error": {
            "description": "Response error",
            "allOf": [
              {
                "$ref": "#/components/schemas/HttpExceptionApiModel"
              }
            ]
          }
        },
        "required": [
          "status",
          "error"
        ]
      },
...
      "ResetPasswordDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "User email",
            "example": "john.doe@mail.com"
          }
        },
        "required": [
          "email"
        ]
      },
...
      "SuccessResponseApiModel": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "description": "Response status",
            "enum": [
              "success"
            ]
          },
          "data": {
            "oneOf": [
              {
                "type": "object",
                "title": "Object",
                "description": "Any JSON",
                "additionalProperties": true
              },
              {
                "type": "array",
                "items": {},
                "title": "Array",
                "description": "Any array"
              },
              {
                "type": "string",
                "title": "String",
                "description": "Any string"
              },
              {
                "type": "number",
                "title": "Number",
                "description": "Any number"
              },
              {
                "type": "boolean",
                "title": "String",
                "description": "True or false"
              }
            ],
            "nullable": true,
            "description": "Response data"
          }
        },
        "required": [
          "status",
          "data"
        ]
      }
    }
  }
}